small fixes

+++ src/main/java/com/example/microserviceexam/dto/userDTO.java
@@ -1,12 +1,11 @@
 package com.example.microserviceexam.dto;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 
 import java.io.Serial;
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -14,9 +13,10 @@ import java.math.BigDecimal;
 public class userDTO implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
+
     private String gymProficiency;
-    private BigDecimal age;
-    private BigDecimal height;
-    private BigDecimal currentWeight;
-    private BigDecimal goalWeight;
+    private Integer age;
+    private Integer height;
+    private Double currentWeight;
+    private Double goalWeight;
 }
